requirements.txt
langchain
langchain-openai
langchain-qdrant
langchain-community
langchain-classic
qdrant-client
python-dotenv
pypdf
fastapi
uvicorn[standard]
python-multipart
httpx
openai
numpy
